Add the ability to add logstash filters
Defining filters in the logstash agent config lets you mutate, delete,
or add fields to log messages. We'll need this for syslog compatibility
because the log messages come through from docker in the field "log"
but the syslog output expects them in the field "message". There are
probably other cases where we'll need this, so exposing it as an
environment variable that can be configured by sweetness.