Adding an implementation of syslog over TLS on TCP
The logstash TCP syslog output doesn't run over TLS. Actually, it doesn't
work at all because it never calls "connect". I guess anybody who's using
this output actually uses UDP? This patch just adds in the working ruby
file during the Docker build, but there's a PR up[0] for pushing the change
into logstash-contrib as well. If and when it gets there, we won't need
to manually patch in the file.

[0] https://github.com/elasticsearch/logstash-contrib/pull/127