Improving bats tests
We're expecting some output from logstash once it's up and running, but
it can take a while to start up and the build was timing out on quay.
Instead of waiting a fixed amount of time and then checking for the output
we expect, now we're making the tests pass as soon as they see the output
they expect and otherwise timing out after 2 minutes, so that in either
case we won't have a hung build on quay.