Restarting GentlemanJerry when it dies
Our logstash processes are configured to die when their Java heap reaches 64MB. This
usually happens when there's a temporary error keeping logstash from connecting or
sending to an endpoint. We should always try again when this happens, so this patch
continually restarts logstash after a 1 second wait whenever it dies. It also prints
an error to stdout that we can grep for with scout so that we know when this is
happening.