Re-read log file to know if Jerry is up
When multiple writers are accessing the same log, it looks like
the current way we grep the logs doesn't work and sometimes misses Jerry
starting. This re-reads the log file every second until Jerry comes up.

This also writes stderr to the log file, and cats the log file if we
fail, which means:

- We can get the stderr logs while running tests without waiting for the
  timeout (via docker exec)
- We see all the log output if Jerry fails to start in the BATS output.