Ensure message ordering from oldest to newest
Using LPUSH means that the order of each individual buffer bucket is
from newest to oldest, which is the opposite of the order we get via
ZRANGE when we query the buffer map (which orders from lowest score to
highest score, and thus from older to newer since the score is a
timestamp).

Using RPUSH fixes that, and this commit also adds tests to ensure that
behavior is maintained.