Gracefully handle missing fields in load-message
As per @aaw, we can't rely on the message forwarded to Redis by Logstash
to contain all the fields we sent in. In this particular case, we can't
rely on the presence of "app" in the message (we can, however, rely on
the message being valid JSON).

Accordingly, this updates load-message.lua to accept messages that don't
have an app, and route them to a "JUNK" buffer.