Restart GentlemanJerry when it crashes
When we shipped Aptible Logs on v2, I introduced a regression by adding
`-o errexit` to the `run-gentleman-jerry.sh` script, which resulted in
not restarting GentlemanJerry if it exited after an unclean shutdown
(which it sometimes does when it uses too much memory).

Unfortunately, we missed this because our tests were inadvertently doing
a clean shutdown of GentlemanJerry, so it exited with 0, and so it
wasn't causing the script to exit.

This patch fixes the runner script by ignoring errors there, and fixes
the tests by forcing an unclean shutdown.