Install stunnel from source
The stunnel package available in Alpine Edge does not work with the
version of Alpine we use (because Alpine now uses and builds against
LibreSSL, but we're on an older version that uses OpenSSL), so we'd
rather build our own here.

Note that mirror stunnel since stunnel doesn't publish old versions of
its packages, and we don't want to *have* to update stunnel just to be
able to build Gentlemanjerry.