Add log tail support for database logging
This adds log tail support for database logs. This isn't really *needed*
to release database logs (since Log Tails won't receive database logs
until we decide to enable that feature on them), but it ensures
Gentlemanjerry will understand these logs if it receives them.

To maintain backwards compatibility with Dumptruck yet avoid
embarrassing collisions between apps whose handle starts with `database`
and databases themselves, we don't change the routing for app logs in
Redis, but include the string `database:` for database logs (since `:`
is not a valid character in a handle, we can't have collisions)

(there is currently only one app whose handle starts with `database`, so
the collision is really hypothetical here, but it's easy to get ahead of
it, so..!)