Add multiple container size flavors
I'm starting to suspect the small heap size we give Gentlemanjerry is
hurting us in terms of performance and causing Log Drains to spend a lot
of time GC'ing rather than actually processing logs;

This isn't intended as a long-term solution; the goal is to let us
experiment with different Gentlemanjerry sizes easily (simply by
changing the Gentlemanjerry tag for a given stack and restarting Log
Drains).

If it turns out this does solve some of our performance problems, we'll
probably revert to having one single image and expose this environment
variable in API / Sweetness.

Note: I'm removing the `master` tag here. We never actually used it.
It's just that we needed a tag to point `latest` at. We'll bring it back
if we end up going back to a single image.