Add PrintGC
We don't have much visibility into what Logstash is doing with its CPU
time, but we have good reason to suspect that in a lot of cases, we're
actually spending our time in GC and not really working on processing
logs.

Adding logging of GC time should help prove (or disprove..!) that
theory.