Handle the json field.
Filebeat requires that if we do JSON manipulation on a log (which is required in order for us to truncate the log message while still ensuring that the end result is valid JSON) then the resulting field is called JSON, or we just dump everything into the top level. While dumping things into the top level gives us control, it also risks overwriting things unintenionally. This ensures that we overwrite nothing by only moving what we care about.