Setup Fluentd as forwarder in GJ
This setups up the current Fluentd to forward logs as GentlemanJerry.

This is tested to work with:

- Elasticsearch
- Papertrail
- LogDNA
- SumoLogic
- HTTPS
- Syslog

This does **not** work for:

- Tails
- Logentries

It is designed to drop-in with sweetness "as-is" and converts the
Logstash configuration provided by sweetness to Fluentd configuration in
the ERB template. We prefer to use log hosting provider created plugins
when available instead of the generic outputs.