Install custom LogDNA plugin
We need to match fields provided by our current (v2) Log Drains. Unfortunately, we can't match perfectly, and we can't even get close out of the box with the native LogDNA fluentd plugin. This updates the fields output by the fluentd plugin to be a closer match to what we're currently doing. A perfect match is impossible because LogDNA doesn't support arbitrary fields in the top level of the output. Instead I had to shove anything bonus in the 'meta' section